14 finish

+++ controllers/cards.js
@@ -1,69 +1,62 @@
 const Card = require('../models/card');
 
-const {
+const IncorrectDataError = require('../errors/incorrectDataError');
+const NotFoundError = require('../errors/notFoundError');
+const ForbiddenError = require('../errors/forbiddenError');
+
+/* const {
   ERROR_CODE_INCORRECT_DATA,
   ERROR_CODE_DATA_NOT_FOUND,
   ERROR_CODE_DEFAULT,
-} = require('../utils/utils');
+} = require('../utils/utils'); */
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch(() => res.status(ERROR_CODE_DEFAULT).send({
-      message: 'Произошла ошибка',
-    }));
+    .catch((err) => next(err));
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
     .then((cards) => res.send(cards))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(ERROR_CODE_INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные при создании карточки, произошла ошибка',
-        });
-        return;
+        next(new IncorrectDataError('Переданы некорректные данные при создании карточки, ошибка валидации'));
+      } else {
+        next(err);
       }
-      res.status(ERROR_CODE_DEFAULT).send({ message: 'Произошла ошибка' });
     });
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(ERROR_CODE_DATA_NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена' });
-        return;
+        throw new NotFoundError('Карточка с указанным _id не найдена');
+      }
+      if (card.owner.toString() !== req.user._id) {
+        throw new ForbiddenError('Нельзя удалить чужую карточку');
       }
       Card.findByIdAndRemove(req.params.cardId)
         .then(() => res.send({ message: `Карточка ${req.params.cardId} удалена` }))
-        .catch(() => {
-          res.status(ERROR_CODE_DEFAULT).send({
-            message: 'Ошибка',
-          });
-        });
+        .catch((err) => next(err));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_CODE_INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные для удаления карточки, произошла ошибка',
-        });
-        return;
+        next(new IncorrectDataError('Переданы некорректные данные для удаления карточки, произошла ошибка'));
+      } else {
+        next(err);
       }
-      res.status(ERROR_CODE_DEFAULT).send({ message: 'Ошибка' });
     });
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(ERROR_CODE_DATA_NOT_FOUND).send({
-          message: 'Карточка с указанным _id не найдена',
-        });
-        return;
+        throw new NotFoundError('Карточка с указанным _id не найдена');
       }
       Card.findByIdAndUpdate(
         req.params.cardId,
@@ -71,48 +64,33 @@ const likeCard = (req, res) => {
         { new: true },
       )
         .then((newCard) => res.send(newCard))
-        .catch(() => {
-          res.status(ERROR_CODE_DEFAULT).send({
-            message: 'Произошла ошибка',
-          });
-        });
+        .catch((err) => next(err));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_CODE_INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные для постановки/снятии лайка, произошла ошибка',
-        });
-        return;
+        next(new IncorrectDataError('Переданы некорректные данные для постановки/снятии лайка, произошла ошибка'));
+      } else {
+        next(err);
       }
-      res.status(ERROR_CODE_DEFAULT).send({
-        message: 'Произошла ошибка',
-      });
     });
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(ERROR_CODE_DATA_NOT_FOUND).send({
-          message: 'Карточка с указанным _id не найдена',
-        });
-        return;
+        throw new NotFoundError('Карточка с указанным _id не найдена');
       }
       Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
         .then((newCard) => res.send(newCard))
-        .catch(() => {
-          res.status(ERROR_CODE_DEFAULT).send({
-            message: 'Произошла ошибка',
-          });
-        });
+        .catch((err) => next(err));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_CODE_INCORRECT_DATA).send({ message: 'Переданы некорректные данные для постановки/снятии лайка, произошла ошибка' });
-        return;
+        next(new IncorrectDataError('Переданы некорректные данные для постановки/снятии лайка, произошла ошибка'));
+      } else {
+        next(err);
       }
-      res.status(ERROR_CODE_DEFAULT).send({ message: 'Произошла ошибка' });
     });
 };
